+++ packages/react-components/src/components/glow/glow.tsx
@@ -1,6 +1,6 @@
-/* eslint-disable @typescript-eslint/naming-convention */
-
 'use client';
+
+/* eslint-disable @typescript-eslint/naming-convention */
 
 import { TokensTypes } from '@lualtek/tokens/platforms/web';
 import tkns from '@lualtek/tokens/platforms/web/tokens.json';