+++ packages/react-components/src/components/skeleton/skeleton.stories.tsx
@@ -1,5 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
+import { Stack, Switch, Textfield } from '@/components';
+
 import { Skeleton } from './skeleton';
 
 const meta = {
@@ -47,4 +49,24 @@ export const WithGap = {
     count: 4,
     height: '48px',
   },
+} satisfies Story;
+
+export const WithChildren = {
+  args: {
+    loading: true,
+  },
+  render: args => (
+    <Stack rowGap={8} fill={false} hAlign="start">
+      <Skeleton {...args}>
+        <Switch />
+      </Skeleton>
+      <Skeleton {...args}>
+        <Textfield label="Label" />
+      </Skeleton>
+      <Skeleton {...args}>
+        Lorem ipsum, dolor sit amet consectetur adipisicing elit. Nihil molestias minus labore quo officia, sit tenetur
+        est qui exercitationem eaque fugiat accusamus natus dolorum possimus, optio temporibus cum numquam voluptas?
+      </Skeleton>
+    </Stack>
+  ),
 } satisfies Story;