+++ packages/react-components/src/components/textfield/textfield.tsx
@@ -61,14 +61,14 @@ export type TextfieldProps = BaseFieldProps & InputHTMLAttributes<HTMLInputEleme
   /**
    * Show the action button on the right side of the input.
    *
-   * @important This prop prevent icon to be displayed.
+   * This prop prevent icon to be displayed.
    */
   showClearButton?: boolean;
   /**
    * Set the hint message to show when the field is invalid.
    *
    * @defaultValue 'Invalid input'
-   * @important This prop is not visible when the field is not invalid, is disabled, or readnly
+   * This prop is not visible when the field is not invalid, is disabled, or readnly
    */
   hint?: ReactNode;
   /**