+++ packages/themes/src/index.ts
@@ -1,47 +1,40 @@
 /* eslint-disable import/extensions */
 /* eslint-disable @typescript-eslint/naming-convention */
 /* eslint-disable no-console, no-restricted-syntax */
-import { createRequire } from 'node:module';
 import path from 'node:path';
+import { fileURLToPath } from 'node:url';
 
 import fs from 'fs-extra';
 import StyleDictionary from 'style-dictionary';
-import { fileURLToPath } from 'url';
+import type { Config } from 'style-dictionary/types';
 
 import OkLCH from './transformers/oklch.ts';
-
-const require = createRequire(import.meta.url);
-// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-const RawColorTokens = require('@lualtek/tokens/platforms/raw/tokens.json');
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
+type ThemeVariants = typeof THEME_VARIANTS[number];
 const THEME_VARIANTS = ['light', 'dark'] as const;
 
 /**
- * This custom config extends the StyleDictionary.Config to change
- * the `tokens` type sinec we use flat/json to inline the tokens
- */
-interface CustomConfig extends StyleDictionary.Config {
-  tokens: any;
-}
-
-/**
  * Generate config for each theme folder
- * and each variant name (eg: monochrome/dark)
+ * and each variant name (eg: light/dark)
  * */
-const getConfig = (name: string, variant: string): CustomConfig => ({
+const getConfig = (name: string, variant: ThemeVariants): Config => ({
   source: [`./src/themes/${name}/${variant}/*.json`],
-  tokens: {
-    ...RawColorTokens as Record<string, unknown>,
-  },
+  include: [
+    // Inject raw tokens to be used in the transformation/references
+    '../tokens/platforms/raw/tokens.json',
+  ],
   platforms: {
     // Build configuration for the web platform
     web: {
       basePxFontSize: 18,
       buildPath: `platforms/web/${name}/`,
-      transformGroup: 'custom-web',
+      transformGroup: 'css',
+      transforms: [
+        'color/oklch',
+      ],
       files: [
         {
           format: 'css/variables',
@@ -64,7 +57,7 @@ const getConfig = (name: string, variant: string): CustomConfig => ({
 });
 
 /**
- * Get all the folders inside the foldeer `themes` (eg, default, monochrome etc)
+ * Get all the folders inside the foldeer `themes` (eg, default, pro etc)
  */
 const availableThemes = fs.readdirSync(path.join(__dirname, 'themes')).filter(
   file => fs.statSync(path.join(__dirname, 'themes', file)).isDirectory(),
@@ -76,7 +69,8 @@ const availableThemes = fs.readdirSync(path.join(__dirname, 'themes')).filter(
  */
 for (const theme of availableThemes) {
   for (const themeVariant of THEME_VARIANTS) {
-    const SDWithConfig = StyleDictionary.extend(getConfig(theme, themeVariant));
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+    const SDWithConfig = new StyleDictionary(getConfig(theme, themeVariant));
 
     /**
    * Register custom transformers to process token values for
@@ -85,26 +79,10 @@ for (const theme of availableThemes) {
     SDWithConfig.registerTransform(OkLCH);
 
     /**
-   * Add the custom transformers to a new transformGroup `custom-web`
-   * used inside getConfig
-   */
-    SDWithConfig.registerTransformGroup({
-      name: 'custom-web',
-      transforms: [
-        'attribute/cti',
-        'name/cti/kebab',
-        'time/seconds',
-        'content/icon',
-        'size/rem',
-        'color/oklch',
-      ],
-    });
-
-    /**
    * Manually run StyleDictionary for all the configured platforms
    */
     console.clear();
-    SDWithConfig.buildAllPlatforms();
+    await SDWithConfig.hasInitialized;
+    await SDWithConfig.buildAllPlatforms();
   }
 }
-