+++ packages/config/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @lualtek/config
 
+## 14.0.0-next.2
+
+### Patch Changes
+
+- @lualtek/tokens@14.0.0-next.2
+
+## 14.0.0-next.1
+
+### Patch Changes
+
+- @lualtek/tokens@14.0.0-next.1
+
+## 14.0.0-next.0
+
+### Major Changes
+
+- d79dc24: Add oklch color notation to color tokens
+
+### Patch Changes
+
+- Updated dependencies [d79dc24]
+  - @lualtek/tokens@14.0.0-next.0
+
 ## 13.6.2
 
 ### Patch Changes