+++ packages/react-components/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @lualtek/react-components
 
+## 14.0.0-next.2
+
+### Minor Changes
+
+- a7b51b0: Add children to Skeleton for dynamic content loading
+- 52ac63a: Add indeterminate state to radio control
+
+### Patch Changes
+
+- 222eb2d: Fix switch thumb animation
+  - @lualtek/icons@14.0.0-next.2
+
+## 14.0.0-next.1
+
+### Patch Changes
+
+- dc9eab7: Test drawer fix
+- dc9eab7: Fix panel border style
+- 8e794f9: Fix modal
+  - @lualtek/icons@14.0.0-next.1
+
+## 14.0.0-next.0
+
+### Major Changes
+
+- cc8ee44: Remove deprecated Card component
+- d79dc24: Add oklch color notation to color tokens
+
+### Patch Changes
+
+- Updated dependencies [d79dc24]
+  - @lualtek/icons@14.0.0-next.0
+
 ## 13.6.2
 
 ### Patch Changes