+++ packages/themes/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @lualtek/themes
 
+## 14.0.0-next.2
+
+## 14.0.0-next.1
+
+## 14.0.0-next.0
+
+### Major Changes
+
+- d79dc24: Add oklch color notation to color tokens
+
 ## 13.6.2
 
 ## 13.6.1